Add docs for new config, remove redundant config and clean up build script
