+++ webpack/server.base.ts
@@ -17,13 +17,14 @@ const serverBaseConfig: Configuration = {
   entry: {
     server: ['./src'],
   },
+  // Treat node-builtins as externals to reduce bundle size during dev.
+  externalsPresets: { node: true },
   resolve: {
+    // This tells webpack to resolve esm modules before commonjs on the server-side.
+    // @apollo/client stopped being bundled without this when moving from 3.4.x to 3.5.x
     mainFields: ['main', 'module'],
+    // defaults only includes .js, .json and .wasm
     extensions: ['.mjs', '.js', '.jsx', '.json', '.ts', '.tsx'],
-  },
-  node: {
-    __dirname: false,
-    __filename: false,
   },
 };
